requirements.txt
Flask==1.0.2
gevent==1.4.0
pymongo==3.8.0
pyjwt==1.7.1
flask-cors==3.0.7
